+++ custom_components/senseme/binary_sensor.py
@@ -1,44 +1,40 @@
-"""Support for Big Ass Fans SenseME occupancy sensor."""
-import logging
+"""Support for Big Ass Fans SenseME occupancy sensors."""
 
 from aiosenseme import SensemeDevice
+
 from homeassistant.components.binary_sensor import (
-    DEVICE_CLASS_OCCUPANCY,
+    BinarySensorDeviceClass,
     BinarySensorEntity,
 )
-from homeassistant.const import CONF_DEVICE
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
+from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SensemeEntity
-from .const import DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
 
 
-async def async_setup_entry(hass, entry, async_add_entities):
+async def async_setup_entry(
+    hass: HomeAssistant,
+    entry: ConfigEntry,
+    async_add_entities: AddConfigEntryEntitiesCallback,
+) -> None:
     """Set up SenseME occupancy sensors."""
-    device = hass.data[DOMAIN][entry.entry_id][CONF_DEVICE]
+    device = entry.runtime_data
     if device.has_sensor:
         async_add_entities([HASensemeOccupancySensor(device)])
 
 
 class HASensemeOccupancySensor(SensemeEntity, BinarySensorEntity):
-    """Representation of a Big Ass Fans SenseME occupancy sensor."""
+    """Representation of a SenseME occupancy sensor."""
 
-    def __init__(self, device: SensemeDevice):
+    _attr_device_class = BinarySensorDeviceClass.OCCUPANCY
+
+    def __init__(self, device: SensemeDevice) -> None:
         """Initialize the entity."""
         super().__init__(device, f"{device.name} Occupancy")
-
-    @property
-    def unique_id(self):
-        """Return a unique identifier for this sensor."""
-        return f"{self._device.uuid}-SENSOR"
+        self._attr_unique_id = f"{device.uuid}-SENSOR"
 
     @property
     def is_on(self) -> bool:
-        """Return True if sensor is occupied."""
+        """Return whether occupancy is detected."""
         return self._device.motion_detected
-
-    @property
-    def device_class(self) -> str:
-        """Return the device class."""
-        return DEVICE_CLASS_OCCUPANCY