+++ custom_components/senseme/light.py
@@ -1,106 +1,76 @@
-"""Support for Big Ass Fans SenseME light."""
-import logging
+"""Support for Big Ass Fans SenseME lights."""
+
 from typing import Any
 
 from aiosenseme import SensemeDevice
+
 from homeassistant.components.light import (
     ATTR_BRIGHTNESS,
-    ATTR_COLOR_TEMP,
-    SUPPORT_BRIGHTNESS,
-    SUPPORT_COLOR_TEMP,
+    ATTR_COLOR_TEMP_KELVIN,
+    ColorMode,
     LightEntity,
 )
-from homeassistant.const import CONF_DEVICE
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
+from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SensemeEntity
-from .const import DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
 
 
-async def async_setup_entry(hass, entry, async_add_entities):
+async def async_setup_entry(
+    hass: HomeAssistant,
+    entry: ConfigEntry,
+    async_add_entities: AddConfigEntryEntitiesCallback,
+) -> None:
     """Set up SenseME lights."""
-    device = hass.data[DOMAIN][entry.entry_id][CONF_DEVICE]
+    device = entry.runtime_data
     if device.has_light:
         async_add_entities([HASensemeLight(device)])
 
 
 class HASensemeLight(SensemeEntity, LightEntity):
-    """Representation of a Big Ass Fans SenseME light."""
+    """Representation of a SenseME light."""
 
-    def __init__(self, device: SensemeDevice):
+    def __init__(self, device: SensemeDevice) -> None:
         """Initialize the entity."""
-        self._device = device
-        if device.is_light:
-            name = device.name
-        else:
-            name = f"{device.name} Light"
+        name = device.name if device.is_light else f"{device.name} Light"
         super().__init__(device, name)
-        self._supported_features = SUPPORT_BRIGHTNESS
+        self._attr_unique_id = f"{device.uuid}-LIGHT"
         if device.is_light:
-            self._supported_features |= SUPPORT_COLOR_TEMP
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique identifier for this light."""
-        return f"{self._device.uuid}-LIGHT"
+            self._attr_supported_color_modes = {ColorMode.COLOR_TEMP}
+            self._attr_color_mode = ColorMode.COLOR_TEMP
+            self._attr_min_color_temp_kelvin = device.light_color_temp_min
+            self._attr_max_color_temp_kelvin = device.light_color_temp_max
+        else:
+            self._attr_supported_color_modes = {ColorMode.BRIGHTNESS}
+            self._attr_color_mode = ColorMode.BRIGHTNESS
 
     @property
     def is_on(self) -> bool:
-        """Return true if light is on."""
+        """Return whether the light is on."""
         return self._device.light_on
 
     @property
     def brightness(self) -> int:
-        """Return the brightness of the light."""
-        light_brightness = self._device.light_brightness * 16
-        if light_brightness == 256:
-            light_brightness = 255
-        return int(light_brightness)
+        """Return the light brightness on Home Assistant's 0-255 scale."""
+        return min(self._device.light_brightness * 16, 255)
 
     @property
-    def color_temp(self) -> int:
-        """Return the color temp value in mireds."""
+    def color_temp_kelvin(self) -> int | None:
+        """Return the light color temperature in Kelvin."""
         if not self._device.is_light:
             return None
-        color_temp = int(round(1000000.0 / float(self._device.light_color_temp)))
-        return color_temp
-
-    @property
-    def min_mireds(self):
-        """Return the coldest color temp that this light supports."""
-        if not self._device.is_light:
-            return None
-        color_temp = int(round(1000000.0 / float(self._device.light_color_temp_max)))
-        return color_temp
-
-    @property
-    def max_mireds(self):
-        """Return the warmest color temp that this light supports."""
-        if not self._device.is_light:
-            return None
-        color_temp = int(round(1000000.0 / float(self._device.light_color_temp_min)))
-        return color_temp
-
-    @property
-    def supported_features(self) -> int:
-        """Flag supported features."""
-        return self._supported_features
+        return self._device.light_color_temp
 
     async def async_turn_on(self, **kwargs: Any) -> None:
-        """Turn on the light."""
-        brightness = kwargs.get(ATTR_BRIGHTNESS)
-        color_temp = kwargs.get(ATTR_COLOR_TEMP)
-        if color_temp is not None:
-            self._device.light_color_temp = int(round(1000000.0 / float(color_temp)))
-        if brightness is None:
-            # no brightness, just turn the light on
+        """Turn on the light and apply supplied attributes."""
+        if (color_temp := kwargs.get(ATTR_COLOR_TEMP_KELVIN)) is not None:
+            self._device.light_color_temp = color_temp
+
+        if (brightness := kwargs.get(ATTR_BRIGHTNESS)) is None:
             self._device.light_on = True
         else:
-            # set the brightness, which will also turn on/off light
-            if brightness == 255:
-                brightness = 256  # this will end up as 16 which is max
-            self._device.light_brightness = int(brightness / 16)
+            self._device.light_brightness = max(round(brightness / 255 * 16), 1)
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Turn off the light."""